pyproject.toml
[tool.poetry]
name = "telemail"
version = "0.1.0"
description = ""
authors = ["volkan <zerovolkan@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
aiogram = "^3.15.0"
google-api-python-client = "^2.156.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
pymongo = "^4.10.1"
bcrypt = "^4.2.1"
scrypt = "^0.8.27"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
